scripts/requirements.txt
requests==2.31.0
pytrends==4.9.2
google-api-python-client==2.118.0
supabase==2.3.4
python-dotenv==1.0.1
